+++ M2/Macaulay2/e/Makefile.files.in
@@ -109,20 +109,20 @@ INTERFACE = \
 	groebner-computations/sagbi \
 	monomial/monideal-minprimes \
 	basic-mutable-matrices/lapack \
-	LLL \
+	computations/LLL \
 	fractionfreeLU \
-	ntl-interface \
+	computations/ntl-interface \
 	fplll-interface \
 	basic-mutable-matrices/mat \
 	basic-mutable-matrices/dmat \
 	coeffrings \
-	Eschreyer \
+	computations/Eschreyer \
 	groebner-computations/gbweight \
 	rings/qring \
 	rings/polyquotient \
 	rings/poly \
 	debug \
-	comp \
+	computations/comp \
 	groebner-computations/comp-gb \
 	groebner-computations/comp-gb-proxy \
 	groebner-computations/comp-gb-declared \
@@ -140,12 +140,12 @@ INTERFACE = \
 	monomial/montable \
 	monomial/montableZZ \
 	free-modules/schreyer-orders \
-	assprime \
+	computations/assprime \
 	buffer \
 	comb \
-	det \
+	computations/det \
 	rings/frac \
-	unused/franzi-brp \
+	computations/BRP \
 	free-modules/freemod \
 	groebner-computations/gb-default \
 	groebner-computations/gb-homog2 \
@@ -164,7 +164,7 @@ INTERFACE = \
 	monoid \
 	monomial/monomial \
 	rings/polyring \
-	pfaff \
+	computations/pfaff \
 	relem \
 	rings/ring \
 	rings/ringelem \
@@ -181,8 +181,8 @@ INTERFACE = \
 COMMANDS = \
 	interface/m2-types \
 	interface/m2-mem \
-	ntl-internal \
-	ntl-debugio \
+	computations/ntl-internal \
+	computations/ntl-debugio \
 	matrices/matrix-kbasis \
 	matrices/matrix-symm \
 	matrices/matrix-sort \
@@ -206,8 +206,8 @@ COMMANDS = \
 	interface/ring \
 	interface/ringelement \
 	interface/ringmap \
-	unused/franzi-gb \
-	unused/franzi-interface
+	computations/BRP-gb \
+	computations/BRP-interface
 
 NAMES_H = \
 	m2tbb \
