+++ M2/Macaulay2/e/computations/comp.hpp
