+++ M2/Macaulay2/e/computations/det.hpp
