+++ M2/Macaulay2/e/computations/ntl-interface.hpp
