+++ M2/Macaulay2/e/free-modules/freemod.cpp
@@ -6,7 +6,7 @@
 #include "rings/polyring.hpp"
 #include "matrices/matrix.hpp"
 #include "matrices/matrix-con.hpp"
-#include "Eschreyer.hpp"
+#include "computations/Eschreyer.hpp"
 #include "groebner-computations/gbring.hpp"
 
 //////////////////////////////////////////////
