+++ M2/Macaulay2/e/free-modules/schreyer-orders.cpp
@@ -5,7 +5,7 @@
 #include "matrices/matrix.hpp"
 #include "comb.hpp"
 #include "rings/polyring.hpp"
-#include "Eschreyer.hpp"
+#include "computations/Eschreyer.hpp"
 #include "finalize.hpp"
 
 SchreyerOrder *SchreyerOrder::create(const Monoid *M)
