+++ M2/Macaulay2/e/groebner-computations/comp-gb.hpp
@@ -3,7 +3,7 @@
 #ifndef _comp_gb_hpp_
 #define _comp_gb_hpp_
 
-#include "comp.hpp"
+#include "computations/comp.hpp"
 class buffer;
 
 // The following are the return values from s_pair_step,
